Chnged database context

+++ routes/changeProfilePic.js
@@ -2,7 +2,7 @@ var express = require('express');
 var router = express.Router();
 
 var multer = require('multer');
-const upload = multer({ dest: './public/images' });
+const upload = multer({ dest: './public/temp' });
 
 const functions = require('../functions');
 
@@ -12,12 +12,15 @@ router.get('/', function (request, response) {
 
 router.post('/', upload.single('profilePic'), function (request, response) {
     var userId = request.user.id;
-    
-    var img = fs.readFileSync(request.file.path);
+
+    if (!request.file || !request.file.path) {
+        response.redirect("/");
+    } else {
+        var img = fs.readFileSync(request.file.path);
         var Photo = img.toString('base64');
         fs.unlinkSync(request.file.path);
         functions.saveProfilePic(Photo, userId);
-
+    }
     response.redirect("/");
 });
